Initial Commit.
This repository comes from a file that originally resided in my
`Personal` repository. I have started to move some of the more
substantial programs (programs that weren't just tests) into their own
repositories and providing style improvements and added documentation to
them. In the future their may be a few more of these repositories
dedicated to a single C program. Stay tuned for that.

- Mitch (mrf-dot)